refactor(sdk): Move the `RoomLatestEvents*` types in their own module.
This patch moves the `RoomLatestEvents*` types in their own new
`room_latest_events` module.
